# api/admin.openapi.yaml
openapi: 3.1.0
...
info:
  title: Latchway Admin API
  version: 1.0.0
  summary: Canonical control-plane API used by the dashboard and CLI.
  description: |
    Normative draft for administrative bootstrap, tenancy, immutable configuration,
    secrets, users, installations, usage, audit and self-tests. Cookie-authenticated
    mutations also require origin validation and `X-CSRF-Token`; scoped bearer API
    tokens are not subject to browser CSRF but receive the same capability checks.
    Browser sessions receive the CSRF value in both the login/bootstrap response
    header and a Secure, SameSite=Strict, non-HttpOnly `__Host-latchway_csrf`
    cookie so a same-origin console reload can recover without persisting it.
    Every successful or rejected mutation produces a redaction-safe audit event.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://gateway.example.test
security:
  - AdminCookie: []
  - AdminToken: []
...
paths:
...
  /admin/v1/installations:
    get:
      tags: [Installations]
      operationId: listInstallations
      summary: List tenant-scoped installations and normalized trust.
      parameters:
        - $ref: '#/components/parameters/EnvironmentFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Installation page without raw evidence or proofs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstallationPage' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/installations/{installationId}:
    get:
      tags: [Installations]
      operationId: getInstallation
      summary: Inspect one installation.
      parameters:
        - $ref: '#/components/parameters/InstallationID'
      responses:
        '200':
          description: Installation details and normalized trust events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Installation' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/installations/{installationId}/revoke:
    post:
      tags: [Installations]
      operationId: revokeInstallation
      summary: Revoke an installation, refresh tokens and sessions.
      parameters:
        - $ref: '#/components/parameters/InstallationID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: { type: string, maxLength: 100 }
      responses:
        '200':
          description: Revoked installation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Installation' }
        default: { $ref: '#/components/responses/Problem' }
...
  /admin/v1/audit-events:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: List append-only redacted administrative audit events.
      parameters:
        - $ref: '#/components/parameters/OrganizationFilter'
        - $ref: '#/components/parameters/EnvironmentFilterOptional'
        - name: actor_kind
          in: query
          schema: { type: string, enum: [admin_user, admin_api_token, system] }
        - name: actor_id
          in: query
          schema: { type: string, pattern: '^(adm|tok)_[0-7][0-9A-HJKMNPQRSTVWXYZ]{25}$' }
        - name: action
          in: query
          schema: { type: string, pattern: '^[a-z][a-z0-9_.]{0,99}$' }
        - name: resource_type
          in: query
          schema: { type: string, pattern: '^[a-z][a-z0-9_.]{0,63}$' }
        - name: resource_id
          in: query
          schema: { type: string, pattern: '^[a-z][a-z0-9]{1,15}_[0-7][0-9A-HJKMNPQRSTVWXYZ]{25}$' }
        - name: source
          in: query
          description: Descriptive transport source. Console is derived from authenticated session transport; CLI versus API is a bounded API-token client claim and is not authorization evidence.
          schema: { type: string, enum: [console, cli, api, system] }
        - name: reason
          in: query
          description: Exact stable operational reason code; arbitrary free-form text is not stored in audit events.
          schema: { $ref: '#/components/schemas/AuditReason' }
        - name: result
          in: query
          schema: { type: string, enum: [succeeded, denied, failed, indeterminate] }
        - name: start
          in: query
          description: Inclusive UTC occurrence-time bound.
          schema: { type: string, format: date-time }
        - name: end
          in: query
          description: Exclusive UTC occurrence-time bound.
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Audit event page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditPage' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/audit-events/{auditEventId}:
    get:
      tags: [Audit]
      operationId: getAuditEvent
      summary: Return one append-only redaction-safe audit event and its field-level diff.
      parameters:
        - name: auditEventId
          in: path
          required: true
          schema: { type: string, pattern: '^aud_[0-7][0-9A-HJKMNPQRSTVWXYZ]{25}$' }
      responses:
        '200':
          description: Audit event detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEvent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /admin/v1/self-test-schedules:
    get:
      tags: [Self-tests]
      operationId: listSelfTestSchedules
      summary: List bounded self-test schedules for one environment.
      description: Returns stable authorization credential IDs and pinned configuration IDs, never token plaintext, provider secret material, or provider bodies.
      parameters:
        - $ref: '#/components/parameters/EnvironmentFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Redaction-safe schedule page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestSchedulePage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Self-tests]
      operationId: createSelfTestSchedule
      summary: Create a bounded schedule using the authenticating durable API token.
      description: |
        This operation rejects browser sessions. The scoped Admin API bearer token
        authenticating this request is bound by stable token ID and active administrator
        membership; its plaintext is never accepted in the document or persisted. The
        active configuration revision and exact current provider secret record versions
        are pinned at creation. Any later token revocation, configuration replacement,
        or secret rotation disables the schedule instead of silently rebinding it.
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSelfTestScheduleRequest' }
      responses:
        '201':
          description: Persistently authorized schedule created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestSchedule' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/self-test-schedules/{scheduleId}:
    get:
      tags: [Self-tests]
      operationId: getSelfTestSchedule
      summary: Return one redaction-safe self-test schedule.
      parameters:
        - $ref: '#/components/parameters/SelfTestScheduleID'
      responses:
        '200':
          description: Exact persisted schedule bindings and lifecycle state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestSchedule' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Self-tests]
      operationId: disableSelfTestSchedule
      summary: Permanently disable a self-test schedule.
      description: Soft-disables future runs while preserving run and audit evidence.
      parameters:
        - $ref: '#/components/parameters/SelfTestScheduleID'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Disabled schedule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestSchedule' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/self-tests:
    post:
      tags: [Self-tests]
      operationId: startSelfTest
      summary: Start a bounded diagnostic or conformance test.
      description: |
        `local` runs durable PostgreSQL, schema and active-configuration checks.
        `upstream` and `openrouter` resolve one upstream and model from the active
        compiled revision, require an exact trusted-input profile plus active USD
        pricing, prove the complete protocol-specific request set's worst-case
        cost is no greater than the caller's ceiling, and only then use the
        server-held credential through the protected dispatcher. Responses, Chat,
        and Anthropic Messages exercise non-streaming and streaming requests with
        output clamping. Embeddings exercises one non-streaming request with a zero
        generated-output bound and explicit skipped streaming/clamp checks.
        OpenRouter remains Chat-only and additionally verifies key access/credit.
        Every supported protocol verifies usage and body-free error normalization.
        No provider credential, prompt, response, URL, or provider error body is
        returned or written to the durable result.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartSelfTestRequest' }
      responses:
        '202':
          description: Self-test accepted; secrets remain server-side.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestRun' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/v1/self-tests/{selfTestId}:
    get:
      tags: [Self-tests]
      operationId: getSelfTest
      summary: Return a redacted self-test result.
      parameters:
        - $ref: '#/components/parameters/SelfTestID'
      responses:
        '200':
          description: Current or terminal self-test state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestRun' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    AdminCookie:
      type: apiKey
      in: cookie
      name: __Host-latchway_admin
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-admin-api-token
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 2048 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    EnvironmentFilter:
      name: environment_id
      in: query
      required: true
      schema: { $ref: '#/components/schemas/EnvironmentID' }
...
  schemas:
...
    EnvironmentID:
      type: string
      pattern: '^env_[A-Za-z0-9_-]{16,128}$'
...
    InstallationPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/Installation' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...